remove rm from unit to fix build server ensure files exist for others

+++ build/run-test-functional-nonet.sh
@@ -4,7 +4,10 @@ cd $DIR/..
 
 set -e
 
-rm app/logs/test.log
+if [ -f  app/logs/test.log ]; then
+  rm app/logs/test.log
+fi
+
 app/console --env=test doctrine:mongodb:schema:drop
 app/console --env=test doctrine:mongodb:fixtures:load
 app/console --env=test sosure:doctrine:index
